Set up about page. Rename index page to use as HomePage.
