+++ src/page-components/AboutPage/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const AboutPage = () => {
+  return <div></div>;
+};
+
+export default AboutPage;
