+++ src/pages/about.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import AboutPage from '../page-components/AboutPage';
+
+const About = () => {
+  return <AboutPage />;
+};
+
+export default About;
